Eclipse cleanup

+++ src/main/java/com/cburch/logisim/data/Value.java
@@ -294,7 +294,7 @@ public class Value {
 			}
 			return "" + value;
 		} else {
-			return "" + ((long) value & 0xFFFFFFFFL);
+			return "" + (value & 0xFFFFFFFFL);
 		}
 	}
