+++ src/main/java/com/cburch/logisim/util/GraphicsUtil.java
@@ -23,7 +23,7 @@ public class GraphicsUtil {
 	static public void switchToWidth(Graphics g, int width) {
 		if (g instanceof Graphics2D) {
 			Graphics2D g2 = (Graphics2D) g;
-			g2.setStroke(new BasicStroke((float) width));
+			g2.setStroke(new BasicStroke(width));
 		}
 	}
